Moved powerset utility to the utils package
