+++ src/y0/util/combinatorics.py
